refactor(studio): extract HistoryStoreCard into its own component module
Move HistoryStoreCard and its HistoryStoreStatus interface out of
System.tsx into frontend/src/components/HistoryStoreCard.tsx so the
card can be consumed by TeamWorkspaceTab without pulling in the entire
System page. KvBlock stays in System.tsx (CatalogStatusCard uses it).

Unused imports (useMutation, useQueryClient, Users, api, Button,
useToast) removed from System.tsx accordingly.
